Extract getSteps()
