+++ packages/ui/src/common/model/machines/getSteps.ts
@@ -1,0 +1,31 @@
+import { Interpreter, State, StateNode } from 'xstate'
+
+interface Step {
+  title: string
+  type: 'past' | 'active' | 'next'
+  isBaby?: boolean
+}
+
+const getActiveNodeOrder = (state: State<any>) => (activeId: number, stateNode: StateNode) => {
+  if (state.matches(stateNode.path.join('.'))) {
+    return stateNode.order
+  }
+
+  return activeId
+}
+
+export const getSteps = (service: Interpreter<any>): Step[] => {
+  const machine = service.machine
+  const stateNodes = machine.stateIds.map((id) => machine.getStateNodeById(id))
+  const activeNodeOrder = stateNodes.reduce(getActiveNodeOrder(service.state), -1)
+
+  return stateNodes
+    .filter((stateNode) => !!stateNode?.meta?.isStep)
+    .map((stateNode) => {
+      return {
+        title: stateNode?.meta?.stepTitle ?? '',
+        type: stateNode.order === activeNodeOrder ? 'active' : stateNode.order < activeNodeOrder ? 'past' : 'next',
+        ...(stateNode.parent?.meta?.isStep ? { isBaby: true } : undefined),
+      }
+    })
+}
